cruds + detalhes

+++ database/migrations/2024_02_05_215443_create_specialties_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('description');
-            $table->float('value', 8, 2);
+            $table->decimal('value', 8, 2);
             $table->timestamps();
         });
     }
